fixes server implementation for testing
server implementation for testing now handles global test server.  Also restructures how message server start and stop is handled
