+++ mock_client_test.go
@@ -1,0 +1,12 @@
+package courier
+
+import "testing"
+
+func TestNewMockClient(t *testing.T) {
+	_, conn, err := NewMockClient("bufnet", testMessageServer.BufDialer)
+	if err != nil {
+		t.Fatalf("could not create client: %s", err)
+	}
+
+	defer conn.Close()
+}
