+++ server_test.go
@@ -3,7 +3,6 @@ package courier
 import (
 	"context"
 	"fmt"
-	"net"
 	"sync"
 	"testing"
 	"time"
@@ -14,15 +13,7 @@ import (
 	"google.golang.org/grpc/test/bufconn"
 )
 
-const bufSize = 1024 * 1024
-
-var lis *bufconn.Listener
-
-func bufDialer(context.Context, string) (net.Conn, error) {
-	return lis.Dial()
-}
-
-func TestMessageServerStartErrorr_Error(t *testing.T) {
+func TestMessageServerStartError_Error(t *testing.T) {
 	method := "testMethod"
 	err := fmt.Errorf("test error")
 	e := &MessageServerStartError{
@@ -68,20 +59,10 @@ func TestMessageServer_PublishMessage(t *testing.T) {
 		chanMap := newChannelMap()
 		defer chanMap.Close()
 		mchan := chanMap.Add(tc.chanMapSubject)
-		server := NewMessageServer(rchan, chanMap)
 		errchan := make(chan error)
-
-		go startTestServer(errchan, server)
-
-		client, conn, err := NewLocalGRPCClient("bufnet2", bufDialer)
-		if err != nil {
-			t.Fatalf("could not create client: %s", err)
-		}
-		defer conn.Close()
+		ms := NewMessageServer("3000", rchan, chanMap)
 
 		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), time.Duration(time.Second*3))
-			defer cancel()
 			m := proto.PublishMessageRequest{
 				Message: &proto.PublishMessage{
 					Id:      uuid.NewString(),
@@ -90,27 +71,21 @@ func TestMessageServer_PublishMessage(t *testing.T) {
 				},
 			}
 
-			_, err := client.PublishMessage(ctx, &m)
+			_, err := ms.PublishMessage(context.Background(), &m)
 			if err != nil {
 				errchan <- err
 			}
 		}()
 
-		if tc.expectedFailure {
-			select {
-			case <-errchan:
-				continue
-			case <-time.After(time.Second * 3):
-				t.Fatal("PublishMessage didn't send err message in time")
-			}
-		}
-
 		select {
 		case <-mchan:
 		case err := <-errchan:
-			t.Fatalf("expected PublishMessage to pass but it failed: %s", err)
-		case <-time.After(time.Second * 3):
-			t.Fatal("PublishMessage didn't send message in time")
+			if tc.expectedFailure {
+				continue
+			}
+			t.Fatalf("failed serving PublishMessage: %s", err)
+		case <-time.After(time.Second):
+			t.Fatal("didn't receive message on Message Channel back in time")
 		}
 	}
 }
@@ -146,20 +121,10 @@ func TestMessageServer_ResponseMessage(t *testing.T) {
 		chanMap := newChannelMap()
 		defer chanMap.Close()
 		mchan := chanMap.Add(tc.chanMapSubject)
-		server := NewMessageServer(rchan, chanMap)
 		errchan := make(chan error)
-
-		go startTestServer(errchan, server)
-
-		client, conn, err := NewLocalGRPCClient("bufnet1", bufDialer)
-		if err != nil {
-			t.Fatalf("could not create client: %s", err)
-		}
-		defer conn.Close()
+		ms := NewMessageServer("3000", rchan, chanMap)
 
 		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), time.Duration(time.Second*3))
-			defer cancel()
 			m := proto.ResponseMessageRequest{
 				Message: &proto.ResponseMessage{
 					Id:      uuid.NewString(),
@@ -168,27 +133,21 @@ func TestMessageServer_ResponseMessage(t *testing.T) {
 				},
 			}
 
-			_, err := client.ResponseMessage(ctx, &m)
+			_, err := ms.ResponseMessage(context.Background(), &m)
 			if err != nil {
 				errchan <- err
 			}
 		}()
 
-		if tc.expectedFailure {
-			select {
-			case <-errchan:
-				continue
-			case <-time.After(time.Second * 3):
-				t.Fatal("ResponseMessage didn't send err message in time")
-			}
-		}
-
 		select {
 		case <-mchan:
 		case err := <-errchan:
-			t.Fatalf("expected ResponseMessage to pass but it failed: %s", err)
-		case <-time.After(time.Second * 3):
-			t.Fatal("ResponseMessage didn't send message in time")
+			if tc.expectedFailure {
+				continue
+			}
+			t.Fatalf("failed serving ResponseMessage: %s", err)
+		case <-time.After(time.Second):
+			t.Fatal("didn't receive message on Message Channel back in time")
 		}
 	}
 }
@@ -222,51 +181,34 @@ func TestMessageServer_RequestMessage(t *testing.T) {
 	for _, tc := range tests {
 		rchan := make(chan ResponseInfo, 1)
 		chanMap := newChannelMap()
-		defer chanMap.Close()
+		ms := NewMessageServer("3000", rchan, chanMap)
 		mchan := chanMap.Add(tc.chanMapSubject)
-		server := NewMessageServer(rchan, chanMap)
+		defer chanMap.Close()
 		errchan := make(chan error)
 
-		go startTestServer(errchan, server)
-
-		client, conn, err := NewLocalGRPCClient("buf", bufDialer)
-		if err != nil {
-			t.Fatalf("could not create client: %s", err)
-		}
-		defer conn.Close()
-
 		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), time.Duration(time.Second*3))
-			defer cancel()
-			m := proto.RequestMessageRequest{
+			_, err := ms.RequestMessage(context.Background(), &proto.RequestMessageRequest{
 				Message: &proto.RequestMessage{
-					Id:      uuid.NewString(),
+					Id:      "messageId",
+					NodeId:  "nodeId",
 					Subject: tc.clientSubject,
 					Content: []byte("test"),
 				},
-			}
-
-			_, err := client.RequestMessage(ctx, &m)
+			})
 			if err != nil {
 				errchan <- err
 			}
 		}()
 
-		if tc.expectedFailure {
-			select {
-			case <-errchan:
-				continue
-			case <-time.After(time.Second * 3):
-				t.Fatal("RequestMessage didn't send err message in time")
-			}
-		}
-
 		select {
 		case <-mchan:
 		case err := <-errchan:
-			t.Fatalf("expected RequestMessage to pass but it failed: %s", err)
-		case <-time.After(time.Second * 3):
-			t.Fatal("RequestMessage didn't send message in time")
+			if tc.expectedFailure {
+				continue
+			}
+			t.Fatalf("failed serving RequestMessage: %s", err)
+		case <-time.After(time.Second):
+			t.Fatal("didn't receive message on Message Channel back in time")
 		}
 
 		close(rchan)
@@ -277,7 +219,7 @@ func TestMessageServer_RequestMessage(t *testing.T) {
 		}
 
 		if count != 1 {
-			t.Fatalf("expected count to be 1 but got %v", count)
+			t.Fatalf("expected a ResponseInfo to be sent through but got %v ResponseInfos", count)
 		}
 	}
 }
@@ -363,7 +305,6 @@ func TestFanForwardMessages(t *testing.T) {
 		if count != tc.channelCount {
 			t.Fatalf("expected passed messages to be %v but got %v", tc.channelCount, count)
 		}
-
 	}
 }
 
@@ -391,14 +332,31 @@ func TestForwardMessage(t *testing.T) {
 	}
 }
 
-func startTestServer(errchan chan error, m *MessageServer) {
-	lis = bufconn.Listen(bufSize)
-	grpcServer := grpc.NewServer()
+var lis = bufconn.Listen(1024 * 1024)
 
-	proto.RegisterMessageServerServer(grpcServer, m)
+/**************************************************************
+Expected Outcomes:
+- should start a grpc server
+**************************************************************/
+func TestStartCourierServer(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	wg := sync.WaitGroup{}
+	wg.Add(1)
+	server := grpc.NewServer()
+
+	go startCourierServer(ctx, &wg, server, lis, "3005")
+
+	cancel()
+
+	done := make(chan struct{})
 	go func() {
-		if err := grpcServer.Serve(lis); err != nil {
-			errchan <- fmt.Errorf("Server exited with error: %v", err)
-		}
+		wg.Wait()
+		close(done)
 	}()
+
+	select {
+	case <-done:
+	case <-time.After(time.Second * 3):
+		t.Fatal("wait group did not finish in time")
+	}
 }
